refactor(Create Posts): since RunAttestationCommand is already listed as an command, just parse it as an action instead of creating a duplicate class
